feat(agents)!: generalize BeeAgent runners
Remove GraniteBeeAgent as it is now part of BeeAgent.

Ref: #178
