+++ src/internals/types.ts
@@ -15,7 +15,6 @@
  */
 
 import type { Primitive } from "@/internals/helpers/guards.js";
-
 type ExcludeArrayLike<T> = object & Exclude<T, any[]>;
 export type ObjectLike<K extends keyof any = string, V = any> = Record<K, V>;
 export type PlainObject<K extends string = string, V = Primitive> = Exclude<
